ChestAura recode, major inventory management update, NameTags, TargetHUD improvements, new utils, fixes, improvements to other modules, ... (#1521)
* Added NoSlotSet

Prevents server from changing your selected hotbar slot.

https://youtu.be/yyBAdSugBFQ

* Minor code improvements

* Major values/subvalues reformatting + isSupported fixes

* Fixed invalid rotation GCD in AntiFireball, Fucker, Tower

* Added new utils

Added EntityUtils.getHealth(entity, fromScoreboard, absorption)

Moved String?.contains(substrings) to StringUtils from GameDetector and made it public.

Various extensions to packets, to convert their x, y, z values to real ones.

Added MovementUtils.strafe support for MoveEvent.

* Simplified Fly Vanilla and NoClip with new utils

* Improved NameTags, TargetHUD element

NameTags:
* renamed values so that they are more clear
* fixed no spaces between player name and ping, ...
* improved scoreboard health logic

TargetHUD:
* implemented scoreboard health detection

* Fixed modules sending wrong ItemStack in C08s

Fixes:
* Fucker, Ignite, Tower

Added extensions:
* mc.thePlayer.onPlayerRightClick()
* mc.thePlayer.sendUseItem()

* Improved Backtrack

Fixed broken sound effect detection.
Implemented new utils.
Cleaned up code.
Refactored confusing variables.

* Major inventory management update

Added smart RepairEquipment to InventoryCleaner:
* repairs stacks by merging them
* checks if useful stack wouldn't lose any enchantment, except if a stack has unbreaking and repaired durability is better than the total expected unbreaking durability
* AutoArmor only equips items that cannot be repaired by InventoryCleaner, essentially waits for them to be repaired first
* repairing output can be sorted straight to hotbar slots and also supports AutoArmor SmartSwap

Fixed stack merging trying to merge different colours of Wool, items with different names, ...

Fixed ArmorComparator rounding making 3x protection 1 as good as 4x protection 1, doesn't round anymore.
* details: https://www.guilded.gg/CCBlueX/groups/1dgpg8Jz/channels/034be45e-1b72-4d5a-bee7-d6ba52ba1657/chat?messageId=c0d88f1e-5ad6-48f3-8acb-d5ab7611164b

Fixed wrong expected unbreaking durability for armor pieces.
Fixed some actions ignoring MinItemAge value.
Fixed sequenced clicks in a row not waiting for NoMove or checking if they should even click.

Improved code.

* ChestAura recode

Now searches for tile entities.

Searches for closest spot to click like KillAura.

Options to ignore already looted chests, to detect chest refills, minimal distance from opponent to open chests.

Optimised to use squared distances where possible.

Debugging value OpenInfo to print when anyone opens chest and how long it took him to take items when he closes it.

Added InventoryUtils.serverOpenContainer.

* Prevent desync in InventoryUtils

* Privatisation of serverOpenContainer

+++ src/main/java/net/ccbluex/liquidbounce/injection/forge/mixins/entity/MixinEntity.java
@@ -200,8 +200,8 @@ public abstract class MixinEntity {
             prevRotationYaw = rotationYaw;
             prevRotationPitch = rotationPitch;
 
-            rotationYaw += yaw * 0.15;
-            rotationPitch -= pitch * 0.15;
+            rotationYaw += (float) (yaw * 0.15);
+            rotationPitch -= (float) (pitch * 0.15);
         }
     }
